Fix: Today Page Not Updating #126
The issue was that the DOM elements were being updated every second, but the data they displayed was not changing because it was only being set on page load or reload. To resolve this, I modified the code so that the today variable, which holds the current date and time, is refreshed every second within the setInterval function. This ensures that the data is current when it’s used to update the DOM elements. Additionally, I consolidated the getAndSetDomElements and setDomElements functions into a single setDomElements function to streamline the code. Now, this unified function is called within the setInterval method, which not only simplifies the code but also guarantees that both the data and the DOM elements are updated in sync every second.